test: organize tests
